init utf8 font support

+++ src/text_renderer/SimpleFont.h
@@ -5,7 +5,7 @@
 // Minimal font struct used by our TextRenderer to avoid depending on Adafruit GFX.
 typedef struct {
   uint16_t bitmapOffset;  ///< Pointer into font->bitmap
-  uint16_t codepoint;     ///< Unicode codepoint for this glyph
+  uint32_t codepoint;     ///< Unicode codepoint for this glyph
   uint8_t width;          ///< Bitmap dimensions in pixels
   uint8_t height;
   uint8_t xAdvance;  ///< Distance to advance cursor (x axis)
